power: Clean up AsyncFileReader tests.
Update the AsyncFileReader tests to use TestMainLoopRunner
instead of rolling their own polling-based harness.

BUG=chromium:320340
TEST=tests still pass

Reviewed-on: https://chromium-review.googlesource.com/177131
Tested-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>

+++ power_manager/powerd/system/async_file_reader.cc
@@ -20,7 +20,7 @@ namespace {
 
 // Since we don't know the file size in advance, we'll have to read successively
 // larger chunks.  Start with 4 KB and double the chunk size with each new read.
-const int kInitialFileReadSize = 4096;
+const size_t kInitialFileReadSize = 4096;
 
 // How often to poll for the AIO status.
 const int kPollMs = 100;
